rename encoder to decoder
