+++ src/Highlighter/Decoder/DecodedString.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace VStelmakh\UrlHighlight\Highlighter\Encoder;
+namespace VStelmakh\UrlHighlight\Highlighter\Decoder;
 
 /**
  * Result of HTML entity decoding: the decoded text and a mapping back to the original encoded positions.
